feat: put credentials in parameter store (#7)
BREAKING CHANGE: This will no longer be compatible with the direct
method of providing the `client_secret` and `refresh_token`. First
because Terraform will store both in its state file, which is generally
unsafe, and second because the `refresh_token` may in rare cases need
to change, and pushing new code requires far more labor than just
updating a single parameter. Additionally, this simplifies the
integration of an actual OAuth2 flow.